+++ master_gas_module/clock.h
@@ -20,8 +20,8 @@ extern int minuto;
 extern int segundo;
 
 extern const char* ntpServer;
-extern long  gmtOffset_sec;
-extern int   daylightOffset_sec;
+extern int32_t  gmtOffset_sec;
+extern int32_t   daylightOffset_sec;
 extern bool ntpConnected;
 extern bool rtcUpdated;
 extern bool rtc_ready;
